# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "fsGaze API",
    "description": "API for analyzing KerML documents and managing safety analysis data.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/api/safety/safety-tasks": {
        "get": {
            "tags": ["Safety Tasks"],
            "summary": "Get all safety tasks",
            "parameters": [
              {
                "in": "query",
                "name": "status",
                "schema": { 
                  "type": "string",
                  "enum": ["open", "started", "in-review", "finished"]
                },
                "description": "Filter by task status"
              }
            ],
            "responses": {
                "200": { "description": "Successful retrieval" }
            }
        }
    },
...
    "/api/safety/safety-tasks/{uuid}/status": {
        "patch": {
            "tags": ["Safety Tasks"],
            "summary": "Update the status of a safety task",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "requestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UpdateSafetyTaskStatus"
                        }
                    }
                }
            },
            "responses": {
                "200": { "description": "Update successful" },
                "400": { "description": "Bad request" }
            }
        }
    },
...
    "/api/safety/sw-components/{uuid}/failure-modes": {
        "get": {
            "tags": ["Failure Modes"],
            "summary": "Get failure modes for a software component",
            "parameters": [
                { "in": "path", "name": "uuid", "required": true, "schema": { "type": "string" } }
            ],
            "responses": {
                "200": { "description": "Successful retrieval" },
                "404": { "description": "Not found" }
            }
        }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateSafetyTaskStatus": {
          "type": "object",
          "properties": {
              "status": { 
                "type": "string",
                "enum": ["open", "started", "in-review", "finished"],
                "description": "New status for the task",
                "example": "started"
              }
          },
          "required": ["status"]
      },
...
    }
  }
}